Move the logic of whether it is in the building to NucleusBuildingManager
